MAINT: removed symlinks, restructured electrochemistry tools cli
